Game

Description

Documentation

data Game

The type to store a game sessions state

Constructors

Game

Fields

gameId :: Int

ID of the game

solution :: SolutionWord

SolutionWord the game is about

players :: [Player]

all participating player

isRunning :: Bool

wether the game is still running

atTurn :: Player

the player to make the next turn

guesses :: String

all chars tried in the past

newGame

Arguments

:: Int
The ID of this session

-> String
The goal of this instance of Hangman

-> Maybe Game
A valid new game or Nothing

Creates a new game session with an completely unsolved given solutionWord

makeATurn

Arguments

:: Player
The player who wants to make a turn

-> Char
The char to try

-> Game
The game to operate on

-> (Game, Bool)
The updated game

Updates the given game session and transists it to the next state

makeASolve

Arguments

:: Player
The player who wants to solve the game

-> String
The solution to try

-> Game
The game to operate on

-> (Game, Bool)
The updated game

Updates the given game session and transists it to the next state

nextPlayerAlive :: Game -> Player

returns the next alive player that is at turn.