Player

Description

Documentation

maxFaliures :: Int

The maximum amount of failures a player is allowed to do before he has to die

data Player

Data type to bundle Imformation about a player within an game session.

Constructors

Player

Fields

playerId :: Int

ID of the player

secret :: String

The secret key of the player

failures :: Int

The amount of failures the player had

isAlive :: Bool

If the player is still alive

newPlayer

Arguments

:: Int
the id for the player

-> Player
created player

Creates a new Player with the given id. The player is initialized with NO secret and alive

newPlayer 4 >>> (4, "", 0, True)

playersAlive

Arguments

:: [Player]
A set of players

-> [Player]
All players currently allowed to play

returns a list of all players alive

wrongGuess

Arguments

:: Player
player that guessed wrong

-> Player
updated player

Updates the player after an wrong try. Kills him after last failure

getPlayerForId

Arguments

:: [Player]
The set of players to search in

-> Int
The id of the player to fetch

-> String
The secret key of the player

-> Maybe Player
A player or Nothing if id or key did not match

Grants access to a player with given id if the key does match

killPlayer

Arguments

:: Player
currently player who did a terrible mistake

-> Player
Now dead player

Sets the state of an player to dead